requirements.txt
tensorflow==2.16.2
keras==3.12.0
opencv-python==4.11.0.86
numpy==1.26.4
matplotlib==3.10.7
pandas==2.3.3
pillow==12.0.0
streamlit==1.51.0
tqdm==4.67.1
mtcnn==1.0.0
deepface==0.0.95
